backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
pydantic-settings==2.2.1
pydantic==2.7.1
cryptography==42.0.5
numpy==1.26.4
scikit-learn==1.4.2
groq==0.9.0
python-dotenv==1.0.1
sentence-transformers==2.7.0
